refactor : Menu 내 price, MenuPrice 객체로 포장
