+++ src/main/java/kitchenpos/domain/Menu.java
@@ -2,7 +2,6 @@ package kitchenpos.domain;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.Objects;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
@@ -22,7 +21,9 @@ public class Menu {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String name;
-	private BigDecimal price;
+
+	@Embedded
+	private MenuPrice price;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "menu_group_id")
@@ -35,16 +36,9 @@ public class Menu {
 	}
 
 	private Menu(String name, BigDecimal price, MenuGroup menuGroup) {
-		validatePrice(price);
 		this.name = name;
-		this.price = price;
+		this.price = MenuPrice.of(price);
 		this.menuGroup = menuGroup;
-	}
-
-	private void validatePrice(BigDecimal price) {
-		if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-			throw new WrongPriceException("메뉴의 가격이 없거나 0보다 작습니다.");
-		}
 	}
 
 	public static Menu create(String name, BigDecimal price, MenuGroup menuGroup) {
@@ -62,7 +56,7 @@ public class Menu {
 			.reduce(BigDecimal::add)
 			.orElse(BigDecimal.ZERO);
 
-		if (price.compareTo(totalPrice) > 0) {
+		if (price.isPositive(totalPrice)) {
 			throw new WrongPriceException("메뉴의 가격이 상품가격의 총합보다 클 수 없습니다.");
 		}
 	}
@@ -76,7 +70,7 @@ public class Menu {
 	}
 
 	public BigDecimal getPrice() {
-		return price;
+		return price.getPrice();
 	}
 
 	public MenuGroup getMenuGroup() {
